+++ src/main/java/dev/ted/jitterticket/eventsourced/domain/concert/TicketsSold.java
@@ -1,7 +1,7 @@
 package dev.ted.jitterticket.eventsourced.domain.concert;
 
 public record TicketsSold(ConcertId concertId,
-                          Long eventSequence,
+                          Integer eventSequence,
                           int quantity,
                           int totalPaid) implements ConcertEvent {
 }
